pyproject.toml
[tool.poetry]
name = "histaware"
version = "0.1.0"
description = ""
authors = ["Leonardo Vida <lleonardovida@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
jupyter = "^1.0.0"
notebook = "^6.1.3"
jupyter-contrib-nbextensions = "^0.5.1"
ipykernel = "^5.3.4"
pandas = "^1.1.1"
spacy = "^2.3.2"
torch = "^1.6.0"
torchvision = "^0.7.0"
transformers = "^3.0.2"
sentence-transformers = "^0.3.4"
seaborn = "^0.10.1"
jupyter-tabnine = "^1.1.0"
jupyterlab-git = "^0.20.0"
pyarrow = "^1.0.1"
jupyterlab_github = "^2.0.0"
python-didl-lite = "^1.2.4"
defusedxml = "^0.6.0"
untangle = "^1.1.1"
xmltodict = "^0.12.0"
kb = "^0.1.8"
pillow = "^7.2.0"
tqdm = "4.40.2"
hdbscan = "^0.8.26"
numpy = "1.17.4"
pyfiglet = "^0.8.post1"
tabulate = "^0.8.7"
typer = "^0.3.2"
hnswlib = "^0.4.0"
gensim = "^3.8.3"
pyenchant = "^3.1.1"
pyLDAvis = "^2.1.2"
loguru = "^0.5.3"
ipython = "7.10"
fasttext = "^0.9.2"
joblib = "^0.17.0"
Keras = "^2.4.3"
tensorflow = "^2.3.1"
spellchecker = "^0.4"
dask = "^2.30.0"
pandarallel = "^1.5.1"
qgrid = "^1.3.1"
nodejs = "^0.1.1"
npm = "^0.1.1"

[tool.poetry.dev-dependencies]
jupyterlab = "^2.2.6"
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
